+++ src/main/java/com/fabriccommunity/spookytime/client/render/PumpcownEntityRenderer.java
@@ -1,4 +1,4 @@
-package com.fabriccommunity.spookytime.client;
+package com.fabriccommunity.spookytime.client.render;
 
 import com.fabriccommunity.spookytime.SpookyTime;
 import com.fabriccommunity.spookytime.entity.PumpcownEntity;
