+++ src/main/java/com/fabriccommunity/spookytime/common/SpookyCommands.java
@@ -2,9 +2,9 @@ package com.fabriccommunity.spookytime.common;
 
 import static net.minecraft.server.command.CommandManager.literal;
 
-import com.fabriccommunity.spookytime.common.cmd.ContibutorsCommand;
-import com.fabriccommunity.spookytime.common.cmd.SpooktoberCommand;
-import com.fabriccommunity.spookytime.common.cmd.SpookyCommand;
+import com.fabriccommunity.spookytime.command.ContibutorsCommand;
+import com.fabriccommunity.spookytime.command.SpooktoberCommand;
+import com.fabriccommunity.spookytime.command.SpookyCommand;
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.tree.LiteralCommandNode;
 import com.mojang.brigadier.tree.RootCommandNode;
