+++ src/main/java/com/fabriccommunity/spookytime/common/SpookyDimensions.java
@@ -1,7 +1,7 @@
 package com.fabriccommunity.spookytime.common;
 
 import com.fabriccommunity.spookytime.SpookyTime;
-import com.fabriccommunity.spookytime.common.world.SpookyBiomeSource;
+import com.fabriccommunity.spookytime.world.SpookyBiomeSource;
 import com.github.draylar.worldtraveler.api.dimension.DimensionBuilder;
 import com.github.draylar.worldtraveler.api.dimension.EntityPlacerBuilder;
 import com.github.draylar.worldtraveler.api.dimension.FabricDimension;
