+++ src/main/java/com/fabriccommunity/spookytime/util/Contributors.java
@@ -8,10 +8,7 @@ import java.util.Map.Entry;
 import com.google.common.collect.ImmutableList;
 
 import net.minecraft.server.command.ServerCommandSource;
-import net.minecraft.text.ClickEvent;
-import net.minecraft.text.ClickEvent.Action;
 import net.minecraft.text.LiteralText;
-import net.minecraft.text.Style;
 import net.minecraft.text.Text;
 import net.minecraft.text.Texts;
 import net.minecraft.text.TranslatableText;
