+++ src/main/java/com/fabriccommunity/spookytime/world/SpookyBiomeSource.java
@@ -1,4 +1,4 @@
-package com.fabriccommunity.spookytime.common.world;
+package com.fabriccommunity.spookytime.world;
 
 import java.util.Collections;
 import java.util.List;
@@ -6,7 +6,7 @@ import java.util.Random;
 import java.util.Set;
 
 import com.fabriccommunity.spookytime.common.SpookyBiomes;
-import com.fabriccommunity.spookytime.common.world.layer.SpookyBiomeLayers;
+import com.fabriccommunity.spookytime.world.layer.SpookyBiomeLayers;
 import com.google.common.collect.Sets;
 
 import net.minecraft.block.BlockState;
