+++ src/main/resources/assets/spookytime/lang/en_us.json
@@ -9,8 +9,8 @@
     "spookytime.cmd.spooktober": "It is Spooktober!",
     "spookytime.cmd.tillspooktober": "It is %d days till Spooktober",
     
-	"item.spookytime.pumpcown_spawn_egg": "Pumpcown Spawn Egg",
-	"item.spookytime.blaze_skirt": "Blaze Skirt",
+    "item.spookytime.pumpcown_spawn_egg": "Pumpcown Spawn Egg",
+    "item.spookytime.blaze_skirt": "Blaze Skirt",
     
     "block.spookytime.tiny_pumpkin": "§4T§6i§en§ay §4P§6u§em§ap§9k§di§6n",
     "block.spookytime.deceased_grass_block": "Deceased Grass Block",
@@ -18,5 +18,7 @@
     "block.spookytime.tainted_gravel": "Tainted Gravel",
     "block.spookytime.tainted_sand": "Tainted Sand",
     
-    "entity.spookytime.pumpcown": "Pumpcown"
+    "entity.spookytime.pumpcown": "Pumpcown",
+    
+    "biome.spookytime.spooky_forest": "Spooky Forest"
 }
